requirements.txt
fastapi
uvicorn
beautifulsoup4
requests
fpdf
python-dotenv
google-generativeai
crewai
